+++ dvlp-back/src/backend/Modules/FleetManagement/Application/DTOs/Request/LIneModelRequestDto.cs
@@ -6,7 +6,7 @@ public class LIneModelRequestDto
 
     public Guid modelId { get; set; }
 
-    public int? capacity { get; set; }
+    public byte? capacity { get; set; }
 
     public string plate { get; set; }
 }
